Identity: move js/Transaction -> js/Identity
